+++ changelog.md
@@ -998,3 +998,17 @@ Main Game
 Main Game
 
 - Finished japHARD Speed branch!
+
+## ? Feb. 2022
+
+Main Game
+
+- Added music: 7 tracks
+  - "/" by TheTastyPi
+  - "v" by TheTastyPi
+  - "+" by yhvr
+  - "z" by yhvr
+  - ">" by TheTastyPi
+  - "Σ" by TheTastyPi
+  - "." by TheTastyPi
+- Added volume option